--- src/main/java/kitchenpos/menu/domain/MenuProductRepository.java
@@ -1,6 +1,0 @@
-package kitchenpos.menu.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MenuProductRepository extends JpaRepository<MenuProduct, Long> {
-}
