+++ src/main/java/kitchenpos/menu/domain/MenuRepository.java
@@ -1,9 +1,6 @@
 package kitchenpos.menu.domain;
 
-import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MenuRepository extends JpaRepository<Menu, Long> {
-
-    Long countByIdIn(List<Long> ids);
 }
