+++ src/main/java/kitchenpos/product/application/ProductValidatorImpl.java
@@ -2,17 +2,18 @@ package kitchenpos.product.application;
 
 import java.util.List;
 import java.util.Optional;
+import kitchenpos.product.ProductValidator;
 import kitchenpos.product.domain.Price;
 import kitchenpos.product.domain.Product;
 import kitchenpos.product.domain.ProductRepository;
 import org.springframework.stereotype.Service;
 
 @Service
-public class ProductValidationServiceImpl implements ProductValidationService {
+public class ProductValidatorImpl implements ProductValidator {
 
     private final ProductRepository productRepository;
 
-    public ProductValidationServiceImpl(final ProductRepository productRepository) {
+    public ProductValidatorImpl(final ProductRepository productRepository) {
         this.productRepository = productRepository;
     }
